schema: Allow additional modifiers to pass validation
As the comment explains:

Since $pop, $pull, $pullAll and $bit cannot be used to break a schema
(i.e. they can't change the type or add data) adding validation for their
content is not a priority.  Perhaps at some point they'll be validated,
but for now just let them pass.