connection: Fix unit test
pymongo apparently changed the way connections are compared at some point to
include checking the ports for matching.  This breaks the test since we mocked
out __init__ to avoid making a connection in the unit test.  Since we're
actually trying to test that the connection is the exact same one the test
should have been using ``is`` anyway though, and this commit fixes that.