connection: Replace Connection with MongoClient
Old and busted; meet new hotness http://emptysquare.net/blog/pymongos-new-default-safe-writes/
MongoClient is new and shiny. As far as functionality goes, the API is backwards compatible with the old Connection class so this change shouldn't break anyone's code. The major change is that writes are safe by default. If you want to use unsafe writes, set w=0. If you use the safe kwarg in your code, you'll get warnings since it has been deprecated in favor of the w option.